Move Billing/Bandwidth Graphs to common format, extract data generation to central function
