+++ html/includes/graphs/bill/bits.inc.php
@@ -1,5 +1,12 @@
 <?php
 
+$datefrom = date('YmdHis', $vars['from']);
+$dateto   = date('YmdHis', $vars['to']);
+
+$rates = getRates($vars['id'], $datefrom, $dateto);
+
+$ports = dbFetchRows('SELECT * FROM `bill_ports` AS B, `ports` AS P, `devices` AS D WHERE B.bill_id = ? AND P.port_id = B.port_id AND D.device_id = P.device_id', array($vars['id']));
+
 // Generate a list of ports and then call the multi_bits grapher to generate from the list
 $i = 0;
